requirements.txt
streamlit
numpy
matplotlib
pandas
Pillow
plotly
scipy
